♻️ refactor: Consolidate VideoPlaybackManager and PlayerManager to eliminate redundancy
- Merged PlayerManager functionality directly into VideoPlaybackManager
- Removed unnecessary delegation layer and state synchronization overhead
- Eliminated 4 PlayerManager files (440 lines of code)
- Updated tests to remove async simulation workarounds
- Maintained all existing public API for backward compatibility
- Updated comment references from PlayerManager to VideoPlaybackManager

This refactoring simplifies the architecture by removing duplicate code and
unnecessary abstraction layers while preserving all functionality.

Fixes #53
