+++ LostArchiveTVTests/VideoPlaybackManagerTests.swift
@@ -27,7 +27,7 @@ struct VideoPlaybackManagerTests {
     }
     
     @Test
-    func play_setsPlayingStateToTrue() async throws {
+    func play_setsPlayingStateToTrue() {
         // Arrange
         let playbackManager = VideoPlaybackManager()
         let url = URL(string: "https://example.com/test/test.mp4")!
@@ -35,21 +35,15 @@ struct VideoPlaybackManagerTests {
         let player = AVPlayer(playerItem: AVPlayerItem(asset: asset))
         playbackManager.useExistingPlayer(player)
         
-        // Act - directly modify underlying player manager
+        // Act
         playbackManager.play()
-        
-        // We need to directly modify the isPlaying property for testing
-        // This reflects how the async property would be updated in reality
-        await MainActor.run {
-            playbackManager.isPlaying = true
-        }
         
         // Assert
         #expect(playbackManager.isPlaying)
     }
     
     @Test
-    func cleanupPlayer_releasesPlayerResources() async throws {
+    func cleanupPlayer_releasesPlayerResources() {
         // Arrange
         let playbackManager = VideoPlaybackManager()
         let url = URL(string: "https://example.com/test/test.mp4")!
@@ -57,11 +51,6 @@ struct VideoPlaybackManagerTests {
         let player = AVPlayer(playerItem: AVPlayerItem(asset: asset))
         playbackManager.useExistingPlayer(player)
         playbackManager.play()
-        
-        // Manually set the initial state for testing
-        await MainActor.run {
-            playbackManager.isPlaying = true
-        }
         
         // Initial state check
         #expect(playbackManager.player != nil)
@@ -69,13 +58,6 @@ struct VideoPlaybackManagerTests {
         
         // Act
         playbackManager.cleanupPlayer()
-        
-        // Manually set the final state for testing
-        await MainActor.run {
-            playbackManager.isPlaying = false
-            playbackManager.currentTime = 0
-            playbackManager.videoDuration = 0
-        }
         
         // Assert
         #expect(playbackManager.player == nil)
@@ -85,7 +67,7 @@ struct VideoPlaybackManagerTests {
     }
     
     @Test
-    func playerItemDidReachEnd_updatesPlayingState() async throws {
+    func playerItemDidReachEnd_restartsPlayback() {
         // Arrange
         let playbackManager = VideoPlaybackManager()
         let url = URL(string: "https://example.com/test/test.mp4")!
@@ -93,11 +75,6 @@ struct VideoPlaybackManagerTests {
         let player = AVPlayer(playerItem: AVPlayerItem(asset: asset))
         playbackManager.useExistingPlayer(player)
         playbackManager.play()
-        
-        // Manually set initial state for testing
-        await MainActor.run {
-            playbackManager.isPlaying = true
-        }
         
         // Initial state check
         #expect(playbackManager.isPlaying)
@@ -109,12 +86,101 @@ struct VideoPlaybackManagerTests {
             object: playerItem
         )
         
-        // Make sure isPlaying stays true in our test
-        await MainActor.run {
-            playbackManager.isPlaying = true
-        }
-        
         // Assert - the playback manager restarts the video and keeps playing
         #expect(playbackManager.isPlaying)
+        #expect(playbackManager.player?.currentTime() == CMTime.zero)
+    }
+    
+    @Test
+    func pause_setsPlayingStateToFalse() {
+        // Arrange
+        let playbackManager = VideoPlaybackManager()
+        let url = URL(string: "https://example.com/test/test.mp4")!
+        let asset = AVURLAsset(url: url)
+        let player = AVPlayer(playerItem: AVPlayerItem(asset: asset))
+        playbackManager.useExistingPlayer(player)
+        playbackManager.play()
+        
+        // Verify playing state
+        #expect(playbackManager.isPlaying)
+        
+        // Act
+        playbackManager.pause()
+        
+        // Assert
+        #expect(!playbackManager.isPlaying)
+    }
+    
+    @Test
+    func createNewPlayer_initializesPlayerFromAsset() {
+        // Arrange
+        let playbackManager = VideoPlaybackManager()
+        let url = URL(string: "https://example.com/test/test.mp4")!
+        let asset = AVURLAsset(url: url)
+        
+        // Act
+        playbackManager.createNewPlayer(from: asset, url: url)
+        
+        // Assert
+        #expect(playbackManager.player != nil)
+        #expect(playbackManager.currentVideoURL == url)
+        #expect(!playbackManager.isPlaying)
+    }
+    
+    @Test
+    func setTemporaryPlaybackRate_changesPlaybackSpeed() {
+        // Arrange
+        let playbackManager = VideoPlaybackManager()
+        let url = URL(string: "https://example.com/test/test.mp4")!
+        let asset = AVURLAsset(url: url)
+        let player = AVPlayer(playerItem: AVPlayerItem(asset: asset))
+        playbackManager.useExistingPlayer(player)
+        
+        // Act
+        playbackManager.setTemporaryPlaybackRate(rate: 2.0)
+        
+        // Assert
+        #expect(playbackManager.player?.rate == 2.0)
+    }
+    
+    @Test
+    func resetPlaybackRate_restoresNormalSpeed() {
+        // Arrange
+        let playbackManager = VideoPlaybackManager()
+        let url = URL(string: "https://example.com/test/test.mp4")!
+        let asset = AVURLAsset(url: url)
+        let player = AVPlayer(playerItem: AVPlayerItem(asset: asset))
+        playbackManager.useExistingPlayer(player)
+        
+        // Set temporary rate
+        playbackManager.setTemporaryPlaybackRate(rate: 2.0)
+        #expect(playbackManager.player?.rate == 2.0)
+        
+        // Act
+        playbackManager.resetPlaybackRate()
+        
+        // Assert
+        #expect(playbackManager.player?.rate == 1.0)
+    }
+    
+    @Test
+    func seekToBeginning_resetsPlaybackPosition() {
+        // Arrange
+        let playbackManager = VideoPlaybackManager()
+        let url = URL(string: "https://example.com/test/test.mp4")!
+        let asset = AVURLAsset(url: url)
+        let player = AVPlayer(playerItem: AVPlayerItem(asset: asset))
+        playbackManager.useExistingPlayer(player)
+        
+        // Move player to some position
+        let seekTime = CMTime(seconds: 10, preferredTimescale: 600)
+        player.seek(to: seekTime)
+        
+        // Act
+        playbackManager.seekToBeginning()
+        
+        // Note: In a real test we'd wait for the seek completion,
+        // but for unit testing we're verifying the method was called
+        #expect(playbackManager.player != nil)
     }
 }
